SSBWeb/requirements.txt
astroid==2.2.4
backcall==0.1.0
decorator==4.3.2
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.10
jedi==0.13.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prettytable==0.7.2
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
tornado==6.0.1
traitlets==4.3.2
typed-ast==1.3.1
wcwidth==0.1.7
wrapt==1.11.1
